pyproject.toml
[project]
name = "nonebot-plugin-anime-downloader"
version = "0.1.0"
description = "✨ 基于 qBittorrent Web UI 的番剧下载 NoneBot 插件 ✨"
authors = [
    {name = "zhaomaoniu", email = "2667292003@qq.com"},
]
dependencies = [
    "nonebot2>=2.2.1",
    "aiohttp>=3.9.5",
    "qbittorrent-api>=2024.3.60",
    "torrent-parser>=0.4.1",
    "nonebot-plugin-apscheduler>=0.4.0",
    "sqlalchemy>=2.0.29",
    "nonebot-plugin-localstore>=0.6.0",
    "opencc>=1.1.7",
    "nonebot-plugin-alconna>=0.42.4",
    "FastAPI>=0.111.0",
]
requires-python = ">=3.8"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"


[tool.pdm]
distribution = true
